@@ -1,7 +1,7 @@
 {
   "name": "mongoose",
   "description": "Mongoose MongoDB ODM",
-  "version": "8.7.3",
+  "version": "8.8.2",
   "author": "Guillermo Rauch <guillermo@learnboost.com>",
   "keywords": [
     "mongodb",
@@ -21,15 +21,15 @@
   "dependencies": {
     "bson": "^6.7.0",
     "kareem": "2.6.3",
-    "mongodb": "6.9.0",
+    "mongodb": "~6.10.0",
     "mpath": "0.9.0",
     "mquery": "5.0.0",
     "ms": "2.1.3",
     "sift": "17.1.3"
   },
   "devDependencies": {
-    "@babel/core": "7.24.7",
-    "@babel/preset-env": "7.25.4",
+    "@babel/core": "7.26.0",
+    "@babel/preset-env": "7.26.0",
     "@typescript-eslint/eslint-plugin": "^8.4.0",
     "@typescript-eslint/parser": "^8.4.0",
     "acquit": "1.3.0",
@@ -41,7 +41,7 @@
     "broken-link-checker": "^0.7.8",
     "buffer": "^5.6.0",
     "cheerio": "1.0.0",
-    "crypto-browserify": "3.12.0",
+    "crypto-browserify": "3.12.1",
     "dotenv": "16.4.5",
     "dox": "1.0.0",
     "eslint": "8.57.0",
@@ -53,11 +53,11 @@
     "lodash.isequal": "4.5.0",
     "lodash.isequalwith": "4.4.0",
     "markdownlint-cli2": "^0.14.0",
-    "marked": "14.1.2",
+    "marked": "14.1.3",
     "mkdirp": "^3.0.1",
-    "mocha": "10.7.3",
+    "mocha": "10.8.2",
     "moment": "2.30.1",
-    "mongodb-memory-server": "10.0.1",
+    "mongodb-memory-server": "10.1.2",
     "ncp": "^2.0.0",
     "nyc": "15.1.0",
     "pug": "3.0.3",
@@ -65,9 +65,9 @@
     "sinon": "19.0.2",
     "stream-browserify": "3.0.0",
     "tsd": "0.31.2",
-    "typescript": "5.6.2",
-    "uuid": "10.0.0",
-    "webpack": "5.95.0"
+    "typescript": "5.6.3",
+    "uuid": "11.0.2",
+    "webpack": "5.96.1"
   },
   "directories": {
     "lib": "./lib/mongoose"
